test: convert orphaned node:test unit tests to playwright unit specs
constraints-check.test.ts and unicity-agg-field.test.ts lived under
api/test/, used node:test, and were never picked up by playwright's
testDir: './tests' — no npm script or pre-push hook ran them. Convert
both to @playwright/test and move them to tests/features/datasets/ as
*.unit.spec.ts, following the shared-operations.unit.spec.ts pattern.
Delete the now-orphaned api/test/ directory.
